Revert "Home 페이지 data 랜더링 레이아웃 구현"
