+++ src/components/atoms/Button.tsx
@@ -1,6 +1,17 @@
+import React from 'react';
 import styled from '@emotion/styled';
 
-export const ButtonItem = styled.button`
+const Button = (props: any) => {
+  return (
+    <a href={props.name}>
+      <ButtonItem>
+        <span>{props.name}</span>
+      </ButtonItem>
+    </a>
+  );
+};
+
+const ButtonItem = styled.button`
   cursor: pointer;
   width: 100%;
   background: linear-gradient(
@@ -15,3 +26,5 @@ export const ButtonItem = styled.button`
   text-transform: capitalize;
   font-size: 16px;
 `;
+
+export default Button;
