--- src/components/atoms/Header/Header.styled.tsx
@@ -1,19 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Wrapper = styled.div`
-  display: flex;
-  align-items: center;
-  margin: 2rem 0 0.8rem;
-  font-size: 1.4rem;
-  font-weight: 600;
-  text-transform: capitalize;
-`;
-
-export const SvgBox = styled.span`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 35px;
-  height: 35px;
-  margin-right: 10px;
-`;
