--- src/components/atoms/Logo/Logo.styled.tsx
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-
-export const LogoWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
