+++ src/components/atoms/SearchInput.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
+import styled from '@emotion/styled';
 import { css } from '@emotion/react';
-import { InputBanner, Input } from './SearchInput.styled';
 
 const SearchInput = ({ value, onChange }: any) => {
   return (
@@ -20,5 +20,24 @@ const SearchInput = ({ value, onChange }: any) => {
     </div>
   );
 };
+
+const InputBanner = styled.div`
+  position: absolute;
+  height: inherit;
+  line-height: 50px;
+  margin-left: 15px;
+  color: rgb(166, 166, 166);
+`;
+const Input = styled.input`
+  width: 100%;
+  height: 100%;
+  background-color: transparent;
+  border-style: none;
+  padding-left: 15px;
+  z-index: 1;
+  :focus {
+    outline: none;
+  }
+`;
 
 export default SearchInput;
