+++ src/components/atoms/svgs/ClipSvg.tsx
@@ -3,72 +3,78 @@ import { css } from '@emotion/react';
 
 const ClipSvg = () => {
   return (
-    <svg
-      width="28px"
-      height="28px"
-      viewBox="0 0 28 28"
-      version="1.1"
-      xmlns="http://www.w3.org/2000/svg"
+    <div
+      css={css`
+        margin-right: 0.5rem;
+      `}
     >
-      <defs>
-        <linearGradient
-          x1="100%"
-          y1="12.4630348%"
-          x2="0%"
-          y2="87.5369652%"
-          id="linearGradient-1"
-        >
-          <stop stopColor="#00FF99" offset="0%"></stop>
-          <stop stopColor="#00CCFF" offset="100%"></stop>
-        </linearGradient>
-        <linearGradient
-          x1="100%"
-          y1="0%"
-          x2="0%"
-          y2="100%"
-          id="linearGradient-2"
-        >
-          <stop stopColor="#00FF99" offset="0%"></stop>
-          <stop stopColor="#00CCFF" offset="100%"></stop>
-        </linearGradient>
-      </defs>
-      <g
-        id="🏡-Homepage"
-        stroke="none"
-        strokeWidth="1"
-        fill="none"
-        fillRule="evenodd"
+      <svg
+        width="28px"
+        height="28px"
+        viewBox="0 0 28 28"
+        version="1.1"
+        xmlns="http://www.w3.org/2000/svg"
       >
-        <g id="Clip-Icon" transform="translate(-275.000000, -737.000000)">
-          <g id="Video" transform="translate(274.500000, 736.500000)">
-            <g
-              id="Clips-Icon-Updated"
-              transform="translate(0.500000, 0.500000)"
-            >
-              <rect
-                id="Rectangle"
-                fill="#121221"
-                opacity="0"
-                x="0"
-                y="0"
-                width="28"
-                height="28"
-              ></rect>
-              <path
-                d="M14.767664,9.14181599 C14.865653,9.24058865 14.9206373,9.37408367 14.9206373,9.51321631 L14.9206373,19.3531441 C14.9206373,19.644388 14.6845375,19.8804878 14.3932936,19.8804878 C14.2535764,19.8804878 14.1195696,19.8250419 14.0206973,19.7263243 L11.046,16.7562481 L8.45234375,16.7571422 C8.16109984,16.7571422 7.925,16.5210424 7.925,16.2297984 L7.925,11.9036065 C7.925,11.6123626 8.16109984,11.3762627 8.45234375,11.3762627 L11.765,11.3762481 L14.0218932,9.13884587 C14.2286523,8.93372713 14.5625453,8.9350569 14.767664,9.14181599 Z M18.3219429,10.1086532 C19.9306724,11.0299103 20.9428571,12.7289223 20.9428571,14.6016892 C20.9428571,16.4706 19.9348523,18.1666374 18.3314041,19.0892943 C18.0531183,19.2494255 17.6962595,19.1561373 17.5343375,18.8809291 C17.3724156,18.6057208 17.4667471,18.2528085 17.7450328,18.0926772 C18.9932239,17.3744426 19.7769176,16.0558241 19.7769176,14.6016892 C19.7769176,13.1445594 18.9899675,11.8236171 17.737676,11.1064782 C17.4590529,10.9469216 17.363977,10.5942046 17.5253178,10.3186629 C17.6866587,10.0431211 18.0433199,9.94909654 18.3219429,10.1086532 Z M16.7254617,11.624252 C17.7337633,12.0981568 18.3903352,13.1050696 18.3903352,14.2289596 C18.3903352,15.2752118 17.8214565,16.2238135 16.9198157,16.7332641 C16.6402613,16.8912195 16.2841577,16.7951504 16.1244359,16.5186876 C15.9647141,16.2422249 16.0618575,15.8900593 16.3414118,15.7321039 C16.8833291,15.4259065 17.2243956,14.8571802 17.2243956,14.2289596 C17.2243956,13.5543057 16.8305288,12.950275 16.224985,12.6656676 C15.9341902,12.5289932 15.8104899,12.1850678 15.9486927,11.8974888 C16.0868956,11.6099098 16.4346669,11.4875775 16.7254617,11.624252 Z"
-                id="Combined-Shape"
-                fill="url(#linearGradient-1)"
-              ></path>
-              <path
-                d="M20.075,0.982142857 L7.925,0.982142857 C4.09056588,0.982142857 0.982142857,4.09056588 0.982142857,7.925 L0.982142857,20.075 C0.982142857,23.9094341 4.09056588,27.0178571 7.925,27.0178571 L20.075,27.0178571 C23.9094341,27.0178571 27.0178571,23.9094341 27.0178571,20.075 L27.0178571,7.925 C27.0178571,4.09056588 23.9094341,0.982142857 20.075,0.982142857 Z M19.9711538,2.91071429 C22.79782,2.91071429 25.0892857,5.20217998 25.0892857,8.02884615 L25.0892857,19.9711538 C25.0892857,22.79782 22.79782,25.0892857 19.9711538,25.0892857 L8.02884615,25.0892857 C5.20217998,25.0892857 2.91071429,22.79782 2.91071429,19.9711538 L2.91071429,8.02884615 C2.91071429,5.20217998 5.20217998,2.91071429 8.02884615,2.91071429 L19.9711538,2.91071429 Z"
-                id="Combined-Shape"
-                fill="url(#linearGradient-2)"
-              ></path>
+        <defs>
+          <linearGradient
+            x1="100%"
+            y1="12.4630348%"
+            x2="0%"
+            y2="87.5369652%"
+            id="linearGradient-1"
+          >
+            <stop stopColor="#00FF99" offset="0%"></stop>
+            <stop stopColor="#00CCFF" offset="100%"></stop>
+          </linearGradient>
+          <linearGradient
+            x1="100%"
+            y1="0%"
+            x2="0%"
+            y2="100%"
+            id="linearGradient-2"
+          >
+            <stop stopColor="#00FF99" offset="0%"></stop>
+            <stop stopColor="#00CCFF" offset="100%"></stop>
+          </linearGradient>
+        </defs>
+        <g
+          id="🏡-Homepage"
+          stroke="none"
+          stroke-width="1"
+          fill="none"
+          fill-rule="evenodd"
+        >
+          <g id="Clip-Icon" transform="translate(-275.000000, -737.000000)">
+            <g id="Video" transform="translate(274.500000, 736.500000)">
+              <g
+                id="Clips-Icon-Updated"
+                transform="translate(0.500000, 0.500000)"
+              >
+                <rect
+                  id="Rectangle"
+                  fill="#121221"
+                  opacity="0"
+                  x="0"
+                  y="0"
+                  width="28"
+                  height="28"
+                ></rect>
+                <path
+                  d="M14.767664,9.14181599 C14.865653,9.24058865 14.9206373,9.37408367 14.9206373,9.51321631 L14.9206373,19.3531441 C14.9206373,19.644388 14.6845375,19.8804878 14.3932936,19.8804878 C14.2535764,19.8804878 14.1195696,19.8250419 14.0206973,19.7263243 L11.046,16.7562481 L8.45234375,16.7571422 C8.16109984,16.7571422 7.925,16.5210424 7.925,16.2297984 L7.925,11.9036065 C7.925,11.6123626 8.16109984,11.3762627 8.45234375,11.3762627 L11.765,11.3762481 L14.0218932,9.13884587 C14.2286523,8.93372713 14.5625453,8.9350569 14.767664,9.14181599 Z M18.3219429,10.1086532 C19.9306724,11.0299103 20.9428571,12.7289223 20.9428571,14.6016892 C20.9428571,16.4706 19.9348523,18.1666374 18.3314041,19.0892943 C18.0531183,19.2494255 17.6962595,19.1561373 17.5343375,18.8809291 C17.3724156,18.6057208 17.4667471,18.2528085 17.7450328,18.0926772 C18.9932239,17.3744426 19.7769176,16.0558241 19.7769176,14.6016892 C19.7769176,13.1445594 18.9899675,11.8236171 17.737676,11.1064782 C17.4590529,10.9469216 17.363977,10.5942046 17.5253178,10.3186629 C17.6866587,10.0431211 18.0433199,9.94909654 18.3219429,10.1086532 Z M16.7254617,11.624252 C17.7337633,12.0981568 18.3903352,13.1050696 18.3903352,14.2289596 C18.3903352,15.2752118 17.8214565,16.2238135 16.9198157,16.7332641 C16.6402613,16.8912195 16.2841577,16.7951504 16.1244359,16.5186876 C15.9647141,16.2422249 16.0618575,15.8900593 16.3414118,15.7321039 C16.8833291,15.4259065 17.2243956,14.8571802 17.2243956,14.2289596 C17.2243956,13.5543057 16.8305288,12.950275 16.224985,12.6656676 C15.9341902,12.5289932 15.8104899,12.1850678 15.9486927,11.8974888 C16.0868956,11.6099098 16.4346669,11.4875775 16.7254617,11.624252 Z"
+                  id="Combined-Shape"
+                  fill="url(#linearGradient-1)"
+                ></path>
+                <path
+                  d="M20.075,0.982142857 L7.925,0.982142857 C4.09056588,0.982142857 0.982142857,4.09056588 0.982142857,7.925 L0.982142857,20.075 C0.982142857,23.9094341 4.09056588,27.0178571 7.925,27.0178571 L20.075,27.0178571 C23.9094341,27.0178571 27.0178571,23.9094341 27.0178571,20.075 L27.0178571,7.925 C27.0178571,4.09056588 23.9094341,0.982142857 20.075,0.982142857 Z M19.9711538,2.91071429 C22.79782,2.91071429 25.0892857,5.20217998 25.0892857,8.02884615 L25.0892857,19.9711538 C25.0892857,22.79782 22.79782,25.0892857 19.9711538,25.0892857 L8.02884615,25.0892857 C5.20217998,25.0892857 2.91071429,22.79782 2.91071429,19.9711538 L2.91071429,8.02884615 C2.91071429,5.20217998 5.20217998,2.91071429 8.02884615,2.91071429 L19.9711538,2.91071429 Z"
+                  id="Combined-Shape"
+                  fill="url(#linearGradient-2)"
+                ></path>
+              </g>
             </g>
           </g>
         </g>
-      </g>
-    </svg>
+      </svg>
+    </div>
   );
 };
