--- src/components/atoms/svgs/UploadIcon.tsx
@@ -1,41 +1,0 @@
-import React from 'react';
-
-const UploadIcon = () => {
-  return (
-    <>
-      <svg
-        width="50%"
-        height="auto"
-        viewBox="0 0 26 30"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <g id="Designs" fill="none" fill-rule="evenodd">
-          <g
-            id="02---squareGIF-Copy"
-            transform="translate(-88 -769)"
-            fill="#FFFFFF"
-          >
-            <g id="video_icon-copy" transform="translate(88 769)">
-              <path
-                d="M24,28 L24,11 L26,11 L26,28.5 L26,30 L0,30 L0,29 L0,0 L1,0 L16,0 L16,2 L2,2 L2,28 L24,28 Z M18,7.5 L18,6 L18,0 L20,0 L20,4 L23,4 L23,7 L26,7 L26,9 L18,9 L18,7.5 Z"
-                id="Combined-Shape"
-              ></path>
-              <g id="addicon" transform="translate(9 12)">
-                <rect id="Rectangle" x="3" width="2" height="8"></rect>
-                <rect
-                  id="Rectangle-Copy"
-                  transform="rotate(90 4 4)"
-                  x="3"
-                  width="2"
-                  height="8"
-                ></rect>
-              </g>
-            </g>
-          </g>
-        </g>
-      </svg>
-    </>
-  );
-};
-
-export default UploadIcon;
