+++ src/components/modules/Carousel.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Carosel = ({ data }: any) => {
+  return (
+    <>
+      <img src={data} />
+    </>
+  );
+};
+
+export default Carosel;
