--- src/components/modules/Carousel/Carousel.tsx
@@ -1,16 +1,0 @@
-import CardLayer from '../../../layer/CardLayer';
-
-interface CardLayerProps {
-  data: any;
-  cardType: string;
-}
-
-const Carosel = ({ data, cardType }: CardLayerProps) => {
-  return (
-    <>
-      <CardLayer data={data} cardType={cardType} />
-    </>
-  );
-};
-
-export default Carosel;
