--- src/components/modules/SearchBar/SearchBar.styled.tsx
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-export const SearchBarWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  height: 45px;
-  background-color: white;
-  border-radius: 5px;
-`;
