--- src/components/modules/UserButton/UserButton.styled.tsx
@@ -1,20 +1,0 @@
-import styled from '@emotion/styled';
-
-export const UserButtonItem = styled.button`
-  cursor: pointer;
-  width: 100%;
-  padding: 8px 0;
-  white-space: nowrap;
-  color: white;
-  background-color: rgb(62, 62, 62);
-  border-style: none;
-  text-transform: capitalize;
-  font-size: 16px;
-  border-radius: 0 5px 5px 0;
-`;
-export const UserSvg = styled.svg`
-  width: 36px;
-  padding: 0 10px;
-  background-color: rgb(82, 82, 82);
-  border-radius: 5px 0 0 5px;
-`;
