+++ src/components/templates/MenuLinkList.tsx
@@ -1,6 +1,25 @@
+import React from 'react';
 import styled from '@emotion/styled';
+import MenuLinkProps from '../../types/menu';
 
-export const UnOrdered = styled.ul`
+const MenuLinkList = ({ list }: { list: MenuLinkProps[] }) => {
+  return (
+    <>
+      <UnOrdered>
+        {list.map((data, idx) => {
+          const { name } = data;
+          return (
+            <Link key={idx}>
+              <Anchor href={name}>{name}</Anchor>
+            </Link>
+          );
+        })}
+      </UnOrdered>
+    </>
+  );
+};
+
+const UnOrdered = styled.ul`
   display: flex;
   height: 35px;
   background: linear-gradient(
@@ -17,12 +36,12 @@ export const UnOrdered = styled.ul`
   margin-left: 14px;
 `;
 
-export const Link = styled.li`
+const Link = styled.li`
   cursor: pointer;
   border-right: 4px solid black;
   height: 100%;
 `;
-export const Anchor = styled.a`
+const Anchor = styled.a`
   display: inline-block;
   text-align: center;
   width: 100%;
@@ -34,3 +53,5 @@ export const Anchor = styled.a`
   text-decoration: none;
   text-transform: capitalize;
 `;
+
+export default MenuLinkList;
