--- src/components/templates/Navbar/Navbar.styled.tsx
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Buttons = styled.div`
-  display: flex;
-  margin-left: 14px;
-  a {
-    margin-right: 5px;
-  }
-`;
