--- src/components/templates/Sidebar/Sidebar.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import SidebarBox from '../../modules/SidebarBox/SidebarBox';
-
-const Sidebar = ({ data }: any) => {
-  return (
-    <>
-      <SidebarBox data={data} />
-    </>
-  );
-};
-
-export default Sidebar;
