+++ src/config/properties.ts
@@ -1,4 +1,3 @@
 export const PROPERTIES = {
   BASE_URL: 'https://api.giphy.com/v1/',
-  API_KEY: 'RqsM2YqZIBX1z7mAkpHMkyhGQpaCvEiG',
 };
