+++ src/layer/SearchInputLayer.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { VAC, VACInput } from 'react-vac';
-import SearchInput from '../components/atoms/SearchInput/SearchInput';
+import SearchInput from '../components/atoms/SearchInput';
 
 const SearchInputLayer = ({ value, onChange }: any) => {
   const searchInputViewProps = {
